Fix remaining 58 (out of 64) tests.
row_number() is inherently undeterministic because it depends on the
segment on which assigns it. For example, for the following query:

  select row_number() over (partition by b) from foo;

if there are two rows with the same b value, they may get assigned a
different row_number, depending on the segment on which the WindowAgg is
run. This is because foo first needs to re-distributed by b (if it
wasn't already so) and that is inherently non-deterministic. But, if it
is executed on master or the table was pre-distributed by b, then this
becomes more deterministic.