Log error if multiple updates to a row are performed by a query
ORCA does not support multiple updates to a row performed by a query.
The handling to disallow it was removed in the process of merge, so add
it back.

Also, now the parse query tree may contain resjunk columns (see
rewriteTargetListUD), so update ORCA to handle the asserts
appropriately. Resjunk columns like CTID are appended in ORCA anyways,
so for now we can drop them in query to dxl translation as they are
added later anyways.