Hungarian notation removal specific to GPOS and Naucrates libraries
As part of moving away from Hungarian notation in the GPORCA codebase,
the integration points between GPORCA and GPDB in the translator have
been renamed to the new convention used in GPORCA. The libraries
currently updated to the new notation in GPORCA are Naucrates and GPOS.
The new naming convention is a custom version of common C++ naming
conventions. The style guide for this convention can be found in the
GPORCA repository.

Also bump ORCA version to 2.69.0

 Co-authored-by: Shreedhar Hardikar <shardikar@pivotal.io>
 Co-authored-by: Melanie Plageman <mplageman@pivotal.io>
 Co-authored-by: Ekta Khanna <ekhanna@pivotal.io>
 Co-authored-by: Abhijit Subramanya <asubramanya@pivotal.io>
 Co-authored-by: Sambitesh Dash <sdash@pivotal.io><Paste>
 Co-authored-by: Dhanashree Kashid <dkashid@pivotal.io>
 Co-authored-by: Omer Arap <oarap@pivotal.io>