Fix memory leak in eval_part_qual
Whenever `eval_part_qual` is invoked, a new qualList object is allocated
and is executed to evaluate a bool result. Once, the qualList is used,
we should free the memory, else we leak memory and can run into OOM
errors for queries on partitioned tables which uses general predicate to
evaluate the partition rule to select.

Signed-off-by: Abhijit Subramanya <asubramanya@pivotal.io>