Add gp_use_legacy_hashops GUC, to keep using the legacy hashops.
This removes the "cdbhash_legacy_ops" alias. pg_dump will set the new GUC,
instad, to force the use of legacy hashops, when creating a dump for
pg_upgrade. It will also set the GUC on all upgraded databases, so any
new tables created after upgrade will also use the legacy opclasses. It
seems better to be consistent after upgrade, so that collocation between
old and new tables work for example.

The idea is that some time after the upgrade, the admin can reorganize
all tables to use the default opclasses instead. At that point, he should
also clear the GUC on the converted databases. (Or rather, the automated
tool that hasn't been written yet, should do that.)

This addresses Simon Gao's and Ning Yu's review comments, on the removal
of the pg_database.hashmethod field.