Optimize hash table in ORCA algebrization
For queries containing joins of tables with many columns, algebrization
in ORCA can be slow due to hash table collisions. For example, a 10-way
join of tables with 1500 cols each will generate a hash table with 80k+
colrefs. By minimizing these collisions, the total optimization time was reduced for
some queries by ~20%, and the algebrization time was reduced by 50%.

We're planning to make hashmaps grow as they reach some load
factor. However, this is a hot spot that we'd like to fix sooner.

Authored-by: Chris Hajas <chajas@pivotal.io>