Fix qual pushdown through Set Operation with window functions.
Previously when checking if it is safe to push down a qual through
window functions in qual_is_pushdown_safe(), we did not consider the
case with Set Operations. As a result, qual would be incorrectly pushed
down through Set Operations with window functions.

In this patch, we recurse through setOperations tree when check for
window functions. Also remove qual_is_pushdown_safe_set_operation()
since it doesn't work correctly and we do not need it to check for window
functions any more.

Picked from 5x commit: 634c856c91de9