gphostcache: fix leftover
Lookup using gphostcache was a fallback mechanism to fix the hostname if
we created wrong address during for the target instance during
recoverseg. The wrong address can result in hostname resolution failure
and will cause gprecoverseg to fail. So, ensure that if there are no
addresses with -n it's a non standard array