Improve cardinality estimation for joins with non-ndv preserving colref
At the point when we calculate stats for a join predicate, if the
predicate contains a computed CColRef, the CColRef may be created from a
non NDV-preserving expression. This commit adds a ndv_preserving field
in CColRef, and adds check for the NDV-preservness of a CColRef when
calculating join stats.

Co-authored-by: Chris Hajas <chajas@vmware.com>