Don't generate predicate which can't be pushed below CTE
Predicates generated on the computed columns of the CTE can't be
pushed down below the CTE Consumer. Even in the case of cte inlining
ORCA adds a Result node to produce the computed columns, so it can't be
pushed futher in the case of inlining too.

Generating additional predicates appears to be not useful in this case
and due to the additional predicates we may also underestimate the rows
produced after the CTE filters.